AB#106837 resizing edit button to be same size of main heading
