+++ packages/layout/src/components/cards/components/button.module.scss
@@ -62,6 +62,7 @@
 .heading4 {
 	font-size: $font-size-2;
 	font-weight: $font-weight-3;
+	width: 100%;
 }
 
 :export {
