+++ packages/layout/src/components/cards/components/card.module.scss
@@ -13,6 +13,11 @@
 		color: $colors-neutral-6;
 		@include removeMarginBottom;
 	}
+
+	.heading {
+		@include styled-as-h3;
+		font-weight: $font-weight-3;
+	}
 }
 
 .statusMsg {
