--- packages/layout/src/components/cards/components/cardContentHeaderSection.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import { UnderlinedButton } from './button';
-
-interface CardContentSectionHeaderProps {
-	sectionHeaderText: string;
-	send?: Function;
-}
-export const CardContentSectionHeader: React.FC<CardContentSectionHeaderProps> = ({
-	sectionHeaderText,
-	send,
-}) => {
-	return (
-		<UnderlinedButton
-			onClick={() => {
-				send('CANCEL');
-			}}
-			isOpen={true}
-			isEditButton={true}
-		>
-			{sectionHeaderText}
-		</UnderlinedButton>
-	);
-};
-
-export default CardContentSectionHeader;
