+++ packages/layout/src/components/cards/components/cardContentSectionHeader.module.scss
@@ -1,0 +1,6 @@
+@import '@tpr/theming/lib/variables.scss';
+
+.contentHeader {
+	width: 50%;
+	padding-right: $space-4;
+}
